organize components
directories with index.ts entries will make this easier to manage
maybe
