+++ src/js/components/tabs/index.tsx
