+++ src/js/components/vertical-tabs/index.tsx
